Fix dictionary save.
