+++ src/activity/page/RecentActivity.cpp
@@ -28,10 +28,12 @@
 #include "Epub/Page.h"
 #include "Epub/Section.h"
 #include "components/recent/RecentLayouts.h"
+#include "dictionary/DictionaryDefinitionLayout.h"
 #include "images/Star.h"
 #include "state/BookProgress.h"
 #include "state/BookSetting.h"
 #include "state/BookState.h"
+#include "state/SavedDictionaryWords.h"
 #include "state/Session.h"
 #include "state/Statistics.h"
 #include "state/SystemSetting.h"
